pyproject.toml
[project]
name = "uv-pakage"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.0.17",
    "streamlit>=1.46.1",
]
